5-flip_bits

+++ 0x14-bit_manipulation/5-flip_bits.c
@@ -10,7 +10,7 @@
 
 unsigned int flip_bits(unsigned long int n, unsigned long int m)
 {
-	int count = 0, xor = n ^ m;
+	unsigned long int count = 0, xor = n ^ m;
 
 	while (xor)
 	{
